% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_coding_tools}
\alias{calc_freqs_coding_tools}
\title{Availability and knowledge of coding langauges}
\usage{
calc_freqs_coding_tools(data, langs)
}
\arguments{
\item{data}{This is generated using the carsurvey2::data functions.}

\item{langs}{Formatted list of strings. langs object is available in carsurvey2::generate_tables()}
}
\value{
data.frame
}
\description{
Function which creates a table of the frequencies for access and knowledge of coding langauges
}
