% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsurvey2-package.R
\docType{package}
\name{carsurvey2}
\alias{carsurvey2}
\title{carsurvey2}
\description{
Analysis for the Coding in Analysis and Research Survey (CARS) wave 2.
The package produces frequencies and outputs for CARS data.
}
