% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{convert_raw}
\alias{convert_raw}
\title{Convert raw data to data.frame}
\usage{
convert_raw(r)
}
\arguments{
\item{r}{api response object}
}
\value{
response content as a data.frame
}
\description{
Convert raw smartsurvey data to data.frame . Extract contents (raw csv) from smartsurvey API request and convert to data.frame
}
