% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_y_lab}
\alias{create_y_lab}
\title{Create custom Y axis label}
\usage{
create_y_lab(ylab, font_size)
}
\arguments{
\item{ylab}{Y axis label}

\item{font_size}{font size used in the chart. This function will return a slightly larger font.}
}
\value{
list of parameters for plotly annotation
}
\description{
Create a custom y axis label (plotly annotation). This label is placed just above the y axis
and is horizontal, to replace the vertically flipped label produced by default.
}
