% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_streaming}
\alias{enforce_streaming}
\title{Enforce streaming rules}
\usage{
enforce_streaming(data)
}
\arguments{
\item{data}{the data which is the output from carsurvey2::tidy_ingest(data)}
}
\value{
the exported data after streaming rules as a dataframe
}
\description{
Input the data as a data frame and apply each of the streaming rules in turn.
}
