% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_module.R
\name{fit_log_reg}
\alias{fit_log_reg}
\title{fit logistic regression}
\usage{
fit_log_reg(formula, data)
}
\arguments{
\item{formula}{string. R formula as character string.}

\item{data}{data.frame.}
}
\value{
model (gml)
}
\description{
fit logistic regression onto data
}
