% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{format_filter_path}
\alias{format_filter_path}
\title{Format filter path}
\usage{
format_filter_path(path)
}
\arguments{
\item{path}{A string}
}
\value{
String
}
\description{
This function cleans the filter names
}
