% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_scales.R
\name{get_3colour_scale}
\alias{get_3colour_scale}
\title{create 3 colour palette}
\usage{
get_3colour_scale(n)
}
\arguments{
\item{n}{the number of colours needed}
}
\value{
RGB colours
}
\description{
create a list of 3 colours in RGB; orange, blue and green for plotting
}
