% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{merge_data_scientist}
\alias{merge_data_scientist}
\title{Merge data science professions}
\usage{
merge_data_scientist(data)
}
\arguments{
\item{data}{a data frame containing cleaned CARS wave 2 data}
}
\value{
the original data with the merged data science column
}
\description{
Merge data science columns into one column named datasci
}
