% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{save_navbar}
\alias{save_navbar}
\title{Save navigation bar}
\usage{
save_navbar(code, path)
}
\arguments{
\item{code}{html code (string)}

\item{path}{path for saving the navigation bar (excluding file name)}
}
\description{
Saves the site navbar.
}
